refactor: Popover 구현 및 ProjectListViewController 파일분리
